Interpolación del jugador, los enemigos y los disparos
Se han interpolado todos los objetos que se mueven del juego, a excepción de las armas y el fondo.

+++ include/Window.h
@@ -28,8 +28,8 @@ class Window
 
         bool modoDios = false;
 
-        float numUpdates = 15.f;
-        float UPDATE_TICK_TIME = 1000.f/numUpdates;
+        int numUpdates = 15;
+        float UPDATE_TICK_TIME = (float)1000/numUpdates;
 
     protected:
